+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.1.2] - 2025-01-??
+## [0.2.0] - 2025-01-??
 ### Added
 
 ### Changed
 - Downgraded error! to warn! logging - to avoid double error logging problems which the library user will also like log (and likely as error).  Reserve error! for unrecoverable error in xum1541.
+- Changed Communication error to take CommunicationKind - to break out different sorts of errors, in particular the StatusValue, which can be used
+on [`Bus::talk`] and [`Bus::Listen`] to infer the lack of a particular device.
 
 ## [0.1.1] - 2025-01-24
 ### Added